feat: 데이터 페칭용 서버 컴포넌트 분리 (#145)
